test: consolidate unit tests into utestDSMCC
